Link Users with UrlInfo, add role-based access and pagination for URLs

+++ UrlShortener.Core/Services/UrlsService.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
+using UrlShortener.Core.Common;
+using UrlShortener.Core.DTOs.Admin;
 using UrlShortener.Core.DTOs.URLs;
 using UrlShortener.Core.Entities;
 using UrlShortener.Core.Exceptions;
@@ -10,7 +12,7 @@ namespace UrlShortener.Core.Services;
 
 public class UrlsService(IUrlsRepository urlsRepository, IUrlShortenerService urlShortenerService, ILogger<UrlsService> logger) : IUrlsService
 {
-    public async Task<string> GenerateUrlAsync(GenerateUrlRequest addRequest, string schema, string host)
+    public async Task<string> GenerateUrlAsync(GenerateUrlRequest addRequest, string schema, string host, string userEmail)
     {
         logger.LogInformation("Generating a short URL for the original URL: {OriginalUrl}", addRequest.OriginalUrl);
         
@@ -25,7 +27,7 @@ public class UrlsService(IUrlsRepository urlsRepository, IUrlShortenerService ur
         var code = await urlShortenerService.GenerateUniqueCode();
         logger.LogInformation("Generated unique code: {Code} for URL: {OriginalUrl}", code, addRequest.OriginalUrl);
         
-        var entity = addRequest.ToEntity(code, schema, host);
+        var entity = addRequest.ToEntity(code, schema, host, userEmail);
 
         var shortenedUrl = await urlsRepository.AddUrlAsync(entity);
         logger.LogInformation("Successfully added the shortened URL to the repository: {ShortenedUrl}", shortenedUrl);
@@ -51,30 +53,47 @@ public class UrlsService(IUrlsRepository urlsRepository, IUrlShortenerService ur
         return urlInfo.OriginalUrl;
     }
 
-    public async Task<UrlGetResponse> GetUrlByIdAsync(Guid id)
+    public async Task<UrlGetResponse> GetUrlByIdAsync(Guid id, string userEmail, bool isAdmin)
     {
         var url = await urlsRepository.GetUrlByIdAsync(id);
 
         if (url is null)
             throw new NotFoundException($"Resource type: {nameof(UrlInfo)} with id: {id} does not exist");
 
+        if (!isAdmin && url.UserEmail != userEmail)
+            throw new ForbiddenException("This operation is forbidden for you!");
+        
         return url.ToDto();
     }
-
-    public async Task<IEnumerable<UrlGetResponse>> GetUrlsAsync(UrlsGetRequest request)
+    
+    public async Task<PageResult<UrlGetResponse>> GetUrlsAsync(UrlsGetRequest request, string userEmail)
     {
-        var urls = await urlsRepository.GetUrlsAsync(request);
+        var (urls, totalCount) = await urlsRepository.GetUrlsAsync(request, userEmail);
 
-        return urls.Select(x => x.ToDto());
+        var mappedUrls = urls.Select(x=>x.ToDto());
+        
+        return new PageResult<UrlGetResponse>(mappedUrls, totalCount, request.PageSize, request.PageNumber);
     }
 
-    public async Task DeleteUrlAsync(Guid id)
+    public async Task DeleteUrlAsync(Guid id, string userEmail, bool isAdmin)
     {
         var url = await urlsRepository.GetUrlByIdAsync(id);
 
         if (url is null)
             throw new NotFoundException($"Resource type: {nameof(UrlInfo)} with id: {id} does not exist");
 
+        if (!isAdmin && url.UserEmail != userEmail)
+            throw new ForbiddenException("This operation is forbidden for you!");
+        
         await urlsRepository.DeleteUrlAsync(url);
+    }
+
+    public async Task<PageResult<UrlGetResponse>> GetAllUrlsAsync(AdminUrlsGetRequest request)
+    {
+        var (urls, totalCount) = await urlsRepository.GetAllUrlsAsync(request);
+
+        var mappedUrls = urls.Select(x=>x.ToDto());
+
+        return new PageResult<UrlGetResponse>(mappedUrls, totalCount, request.PageSize, request.PageNumber);
     }
 }
